rag/requirements.txt
openai>=1.0.0
supabase>=2.0.0
python-dotenv>=0.19.0
tweepy>=4.12.0 